Ignore PHP errors related to the version being run (#12)
This is an issue that is likely coming from PHPCompatibility or PHPCompatibility-WP. The issue was present in WPCS (but fixed in 3.0). There is no real benefit for leaving error reporting in when it has to do with a CS ruleset, so I think we can safely just ignore errors.
See: https://github.com/WordPress/WordPress-Coding-Standards/issues/2035
See also: https://gist.github.com/shreyasikhar/f4495ac952586096eb81d505e9e914a4